anonymized user data included in MapboxGL issues for support and follow-up
