+++ src/utils/webVitals.js
@@ -1,23 +1,9 @@
 import { onCLS, onFCP, onINP, onLCP, onTTFB } from 'web-vitals';
 import ReactGA4 from 'react-ga4';
 
-import { CLIENT_BUILD_VERSION } from '../constants';
 import getWindowLocation from './getWindowLocation';
-import { hashString } from './string';
 
-export const createUserAnalyticsData = (user = {}, selectedUserProfile = {}, serverVersion = 'unknown') => {
-  const activeUser = selectedUserProfile.id ? selectedUserProfile : user;
 
-  return {
-    user_role: activeUser.role || 'unknown',
-    organization: getWindowLocation().hostname,
-    user_id_hash: hashString(activeUser.id),
-    is_staff: activeUser.is_staff || false,
-    is_superuser: activeUser.is_superuser || false,
-    client_version: CLIENT_BUILD_VERSION,
-    server_version: serverVersion,
-  };
-};
 
 const isLocalhost = () => {
   return getWindowLocation().hostname === 'localhost' || getWindowLocation().hostname === '127.0.0.1';
